View list of collection instruments uploaded
* User story *
As a Collection Exercise Coordinator
I need: to view the collection instruments that the system holds for the
given collection exercise
So that: I am happy that the collection instrument(s) loaded are correct

* Acceptance criteria *
Given the collection instruments have been loaded
When the user navigates to the collection exercise details page
Then they are able to see the filename (surveyid_formtype_period) of all
collection instruments that have been loaded
