+++ response_operations_ui/__init__.py
@@ -26,5 +26,5 @@ logger_initial_config(service_name='response-operations-ui', log_level=app.confi
 
 
 from response_operations_ui import error_handlers  # NOQA # pylint: disable=wrong-import-position
-from response_operations_ui.views import surveys  # NOQA # pylint: disable=wrong-import-position
+from response_operations_ui.views import collection_exercise, surveys  # NOQA # pylint: disable=wrong-import-position
 from response_operations_ui.views import info  # NOQA # pylint: disable=wrong-import-position
