+++ tests/views/test_survey.py
@@ -1,13 +1,11 @@
 import json
 import unittest
-from io import BytesIO
 from unittest.mock import MagicMock
 
-from requests import RequestException
 import requests_mock
+from requests import RequestException
 
 from response_operations_ui import app
-
 
 url_get_survey_list = f'{app.config["BACKSTAGE_API_URL"]}/survey/surveys'
 with open('tests/test_data/survey/survey_list.json') as json_data:
@@ -15,10 +13,6 @@ with open('tests/test_data/survey/survey_list.json') as json_data:
 url_get_survey_by_short_name = f'{app.config["BACKSTAGE_API_URL"]}/survey/shortname/bres'
 with open('tests/test_data/survey/survey.json') as json_data:
     survey_info = json.load(json_data)
-url_get_collection_exercise = f'{app.config["BACKSTAGE_API_URL"]}/collection-exercise/test/000000'
-with open('tests/test_data/collection_exercise/collection_exercise_details.json') as json_data:
-    collection_exercise_details = json.load(json_data)
-url_upload_collection_instrument = f'{app.config["BACKSTAGE_API_URL"]}/collection-instrument/test/000000'
 
 
 class TestSurvey(unittest.TestCase):
@@ -79,69 +73,3 @@ class TestSurvey(unittest.TestCase):
 
         self.assertEqual(response.status_code, 200)
         self.assertIn("FAIL".encode(), response.data)
-
-    @requests_mock.mock()
-    def test_collection_exercise_view(self, mock_request):
-        mock_request.get(url_get_collection_exercise, json=collection_exercise_details)
-
-        response = self.app.get("/surveys/test/000000")
-
-        self.assertEqual(response.status_code, 200)
-        self.assertIn("Business Register and Employment Survey".encode(), response.data)
-        self.assertIn("221_201712".encode(), response.data)
-
-    @requests_mock.mock()
-    def test_collection_exercise_view_fail(self, mock_request):
-        mock_request.get(url_get_collection_exercise, status_code=500)
-
-        response = self.app.get("/surveys/test/000000")
-
-        self.assertEqual(response.status_code, 200)
-        self.assertIn("FAIL".encode(), response.data)
-
-    @requests_mock.mock()
-    def test_upload_collection_instrument(self, mock_request):
-        file = dict(
-            ciFile=(BytesIO(b'data'), 'test.xlsx'),
-        )
-        mock_request.post(url_upload_collection_instrument, status_code=201)
-        mock_request.get(url_get_collection_exercise, json=collection_exercise_details)
-
-        response = self.app.post("/surveys/test/000000", data=file)
-
-        self.assertEqual(response.status_code, 200)
-        self.assertIn("Collection instrument loaded".encode(), response.data)
-
-    @requests_mock.mock()
-    def test_failed_upload_collection_instrument(self, mock_request):
-        file = dict(
-            ciFile=(BytesIO(b'data'), 'test.xlsx'),
-        )
-        mock_request.post(url_upload_collection_instrument, status_code=500)
-        mock_request.get(url_get_collection_exercise, json=collection_exercise_details)
-
-        response = self.app.post("/surveys/test/000000", data=file)
-
-        self.assertEqual(response.status_code, 200)
-        self.assertIn("FAIL".encode(), response.data)
-
-    @requests_mock.mock()
-    def test_no_upload_collection_instrument_when_bad_extension(self, mock_request):
-        file = dict(
-            ciFile=(BytesIO(b'data'), 'test.html'),
-        )
-        mock_request.get(url_get_collection_exercise, json=collection_exercise_details)
-
-        response = self.app.post("/surveys/test/000000", data=file)
-
-        self.assertEqual(response.status_code, 200)
-        self.assertNotIn("Collection instrument loaded".encode(), response.data)
-
-    @requests_mock.mock()
-    def test_no_upload_collection_instrument_when_no_file(self, mock_request):
-        mock_request.get(url_get_collection_exercise, json=collection_exercise_details)
-
-        response = self.app.post("/surveys/test/000000")
-
-        self.assertEqual(response.status_code, 200)
-        self.assertNotIn("Collection instrument loaded".encode(), response.data)
